+++ src/main/java/com/telepathicgrunt/the_bumblezone/capabilities/CapabilityEntityPosAndDim.java
@@ -9,27 +9,27 @@ import net.minecraftforge.common.capabilities.CapabilityManager;
 
 import javax.annotation.Nullable;
 
-public class CapabilityPlayerPosAndDim {
+public class CapabilityEntityPosAndDim {
 		//the capability itself
-		@CapabilityInject(IPlayerPosAndDim.class)
-		public static Capability<IPlayerPosAndDim> PAST_POS_AND_DIM = null;
+		@CapabilityInject(IEntityPosAndDim.class)
+		public static Capability<IEntityPosAndDim> PAST_POS_AND_DIM = null;
 
 		//registers the capability and defines how it will read/write data from nbt
 		public static void register() {
-			CapabilityManager.INSTANCE.register(IPlayerPosAndDim.class, new Capability.IStorage<IPlayerPosAndDim>() 
+			CapabilityManager.INSTANCE.register(IEntityPosAndDim.class, new Capability.IStorage<IEntityPosAndDim>()
 			{
 				@Override
 				@Nullable
-				public INBT writeNBT(Capability<IPlayerPosAndDim> capability, IPlayerPosAndDim instance, Direction side) 
+				public INBT writeNBT(Capability<IEntityPosAndDim> capability, IEntityPosAndDim instance, Direction side)
 				{
 					return instance.saveNBTData();
 				}
 
 				@Override
-				public void readNBT(Capability<IPlayerPosAndDim> capability, IPlayerPosAndDim instance, Direction side, INBT nbt) 
+				public void readNBT(Capability<IEntityPosAndDim> capability, IEntityPosAndDim instance, Direction side, INBT nbt)
 				{
 					instance.loadNBTData((CompoundNBT) nbt);
 				}
-			}, PlayerPositionAndDimension::new);
+			}, EntityPositionAndDimension::new);
 		}
 }