+++ src/main/java/com/telepathicgrunt/the_bumblezone/capabilities/PastPosAndDimProvider.java
@@ -11,11 +11,11 @@ import net.minecraftforge.common.util.LazyOptional;
 public class PastPosAndDimProvider implements ICapabilityProvider, ICapabilitySerializable<CompoundNBT> {
 
 	//the capability itself
-	@CapabilityInject(IPlayerPosAndDim.class)
-	public static Capability<IPlayerPosAndDim> PAST_POS_AND_DIM = null;
+	@CapabilityInject(IEntityPosAndDim.class)
+	public static Capability<IEntityPosAndDim> PAST_POS_AND_DIM = null;
 	
 	//The instance of the capability? I think?
-	private PlayerPositionAndDimension instance = (PlayerPositionAndDimension) PAST_POS_AND_DIM.getDefaultInstance();
+	private EntityPositionAndDimension instance = (EntityPositionAndDimension) PAST_POS_AND_DIM.getDefaultInstance();
 
 
 	//returns the capability attached to player
@@ -23,7 +23,7 @@ public class PastPosAndDimProvider implements ICapabilityProvider, ICapabilitySe
 	public <T> LazyOptional<T> getCapability(Capability<T> cap, Direction side) {
 		if (cap == PAST_POS_AND_DIM) {
 			if(instance == null) {
-				instance = new PlayerPositionAndDimension();
+				instance = new EntityPositionAndDimension();
 			}
 			
 			return LazyOptional.of(() -> instance).cast();